move `Column` into it's own file
